more convenient run{Infer,UnifyTy} API
Previously runInfer took both an Infer action and a follow-up
GeneralizeTy final action. It was awkward to call runInfer because the
Infer argument is usually a do block while the GeneralizeTy argument is
usually a one-liner, thus causing the do block to be wrapped in
parenthesis. The new API uses a single do block containing the Infer
action and ending with `pure $ do` and the GeneralizeTy action, which
is more aestetically pleasing.

Another advantage is that the different branches of the Infer
action can each have their own GeneralizeTy final action (see the
implementation of runInfer in terms of runUnifyTy for example). This was
not possible with the previous API.