delete UnifyExtensible
there is no need for this complicated unification-inside-a-unification:
the only case in which two UVarExtensibles were unified was when two
contexts were unified, and in that case they get the same PContext (and
thus the same 'moveVarsAllowed'). So it sufficies to store a boolean,
there is no need to jump through an extra layer of indirection in order
to get that boolean.