revamp MonadUnifyCtx
Previously, it only had a liftUnifyCtx method, like the other MonadFoo
typeclasses in this codebase. That makes sense for monad transformers
which are embedded in a stack of monad transformers and the MonadFoo
instance just calls 'lift' the correct number of times.

For MonadUnifyCtx, however, I want UnifyCtxWithConstraints to have its
own implementation for 'insert' and 'close', one which applies the
constraints in question. So it makes more sense for MonadUnifyCtx to
list all the individual operations which are supported by MonadUnifyCtx.