use variable names matching the MkNode numbering
the test results reveal the numbers used in the internal representation
of the UVarTy's. since the variable names were using a 1-based numbering
scheme and the internal representation was using a 0-based numbering
scheme, it was difficult to understand the test results. Now they both
use a 0-based numbering scheme.